requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.1.0
xlsxwriter>=3.1.0
openpyxl>=3.1.0
python-docx>=0.8.11
pypdf>=3.17.0
